Refactor product css soo that single css server all products in different pages
