+++ App/src/components/electronicsProducts/allProd.module.css
@@ -1,9 +1,9 @@
-.all__electronics__products {
-  border: 2px solid green;
+.all__products {
+  /* border: 2px solid green; */
   width: 100%;
 }
 
-.all__elect__products__display {
+.all__products__display {
   display: grid;
   grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
   gap: 20px;
@@ -15,12 +15,12 @@
 }
 
 @media screen and (max-width: 768px) {
-  .all__electronics__products {
+  .all__products {
     /* width: 80%; */
     /* margin-left: auto; */
   }
 
-  .all__elect__products__display {
+  .all__products__display {
     /* grid-template-columns: 1fr;
     max-width: 200px; */
   }
@@ -34,7 +34,7 @@
 }
 
 @media screen and (max-width: 650px) {
-  /* .all__electronics__products {
+  /* .all__products {
     width: 100%;
   } */
 }
