--- App/src/components/electronicsProducts/sinlgeElectProduct.module.css
@@ -1,47 +1,0 @@
-.single__electronics__product {
-  /* border: 2px solid red; */
-  position: relative;
-}
-
-.single__electronics__product > img {
-  width: 100%;
-  height: 200px;
-}
-
-.favourite__icon {
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  color: #fff;
-  font-size: 25px;
-  cursor: pointer;
-  transition: ease 0.5s all;
-}
-.favourite__icon:hover {
-  transform: scale(1.2);
-}
-
-.short__desc {
-  display: flex;
-  justify-content: space-between;
-  margin-block: 10px;
-  opacity: 0.7;
-}
-
-.short__desc h3:first-child {
-  width: 80%;
-}
-
-.short__desc h3:last-child {
-  /* border: 2px solid green; */
-  display: flex;
-}
-
-/* .short__desc h3 span:first-child,
-.short__desc h3 span:last-child {
-  font-size: 13px;
-} */
-
-.cart__icon {
-  /* color: #fff; */
-}
