+++ App/src/components/electronicsProducts/sinlgeProduct.module.css
@@ -1,9 +1,9 @@
-.single__fashion__product {
+.single__product {
   /* border: 2px solid red; */
   position: relative;
 }
 
-.single__fashion__product > img {
+.single__product > img {
   width: 100%;
   height: 200px;
 }
@@ -21,26 +21,26 @@
   transform: scale(1.2);
 }
 
-.fashion__short__desc {
+.short__desc {
   display: flex;
   justify-content: space-between;
-  align-items: center;
   margin-block: 10px;
   opacity: 0.7;
 }
-.fashion__short__desc h4:first-child {
+
+.short__desc h3:first-child {
   width: 80%;
 }
 
-.fashion__short__desc h3:last-child {
+.short__desc h3:last-child {
   /* border: 2px solid green; */
   display: flex;
 }
 
-.fashion__short__desc h3 span:first-child,
-.fashion__short__desc h3 span:last-child {
+/* .short__desc h3 span:first-child,
+.short__desc h3 span:last-child {
   font-size: 13px;
-}
+} */
 
 .cart__icon {
   /* color: #fff; */
