--- App/src/components/fashionProducts/allFashion.module.css
@@ -1,13 +1,0 @@
-.all__fashion__products {
-  /* border: 2px solid green; */
-  width: 100%;
-  margin-left: 3rem;
-}
-
-.all__fashion__products__display {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-  gap: 20px;
-  margin-top: 20px;
-  /* justify-items: center; */
-}
