+++ App/src/components/fashionProducts/singleFashion.tsx
@@ -1,8 +1,7 @@
 import React from "react";
-import { cartIcon } from "@/assets/icons";
 import { FaHeart, FaCartPlus } from "react-icons/fa";
 import Image from "next/image";
-import styles from "./singleFashion.module.css";
+import styles from "../electronicsProducts/sinlgeProduct.module.css";
 import Button from "../general/button/button";
 
 interface FashionProps {
@@ -12,11 +11,11 @@ interface FashionProps {
 
 const SingleFashionProduct = ({ image, category }: FashionProps) => {
   return (
-    <section className={styles.single__fashion__product}>
+    <section className={styles.single__product}>
       <Image src={image} alt={category} width={200} height={100} />
       <FaHeart title="Add to favourite" className={styles.favourite__icon} />
 
-      <div className={styles.fashion__short__desc}>
+      <div className={styles.short__desc}>
         <h3>Red Shimmer Summer Dress</h3>
         <h3>
           <span>$</span>
