+++ App/src/components/sideBar/sideBar.module.css
@@ -15,6 +15,7 @@
     background-color: #fff;
     padding-top: 9rem;
     transition: ease 0.5s all;
+    border-right: 2px solid var(--solidFadedBlue);
   }
 
   .remove__sidebar {
@@ -22,7 +23,7 @@
   }
 
   .add__sidebar {
-    left: 0;
+    left: var(--small);
   }
 
   .close__sidebar {
